requirements.txt
aniso8601==7.0.0
asn1crypto==0.24.0
bcrypt==3.1.7
cffi==1.12.3
Click==7.0
cryptography==2.7
Flask==1.1.1
Flask-GraphQL==2.0.0
Flask-Cors==3.0.8
Flask-PyMongo==2.3.0
graphene==2.1.8
graphql-core==2.2.1
graphql-relay==2.0.0
graphql-server-core==1.1.1
gunicorn==19.9.0
itsdangerous==1.1.0
Jinja2==2.10.1
jwcrypto==0.6.0
MarkupSafe==1.1.1
promise==2.2.1
pycparser==2.19
PyJWT==1.7.1
pymongo==3.9.0
python-dotenv==0.10.3
python-jwt==3.2.4
Rx==1.6.1
six==1.12.0
Werkzeug==0.16.0
